update current version

+++ bookDelBook.c
@@ -6,6 +6,9 @@
  *  Log:
  *      11-Jan-2022 start by copying bookDelGenre.c
  *      15-Sep-2022 add Access-Control-Allow-Origin: * CORS http header
+ *      11-Oct-2022 update commments
+ *      11=Oct-2022 use EXIT_SUCCESS and EXIT_FAILURE on returns
+ *      11-Oct-2022 validate QUERY_STRING is not NULL or empty
  *  Enhancements:
 */
 
@@ -21,12 +24,12 @@
 
 #define MAXLEN 1024
 
-// global declarations
+// global declarations -------------------------------------------------------------------------------------------------
 
-char *sgServer = "192.168.0.13";                                                               //mysqlServer IP address
-char *sgUsername = "gjarman";                                                              // mysqlSerer logon username
-char *sgPassword = "Mpa4egu$";                                                    // password to connect to mysqlserver
-char *sgDatabase = "risingfast";                                                // default database name on mysqlserver
+char *sgServer = "192.168.0.13";                                                                //mysqlServer IP address
+char *sgUsername = "gjarman";                                                               // mysqlSerer logon username
+char *sgPassword = "Mpa4egu$";                                                     // password to connect to mysqlserver
+char *sgDatabase = "risingfast";                                                 // default database name on mysqlserver
 
 MYSQL *conn;
 MYSQL_RES *res;
@@ -41,12 +44,12 @@ int main(void) {
 
     char caSQL[SQL_LEN] = {'\0'};
 
-// print the html content type and CORS <header> block ----------------------------------------------------------------
+// print the html content type and CORS <header> block -----------------------------------------------------------------
 
     printf("Content-type: text/html\n");
     printf("Access-Control-Allow-Origin: *\n\n");
 
-// Initialize a connection and connect to the database
+// Initialize a connection and connect to the database                                                                  
 
     conn = mysql_init(NULL);
 
@@ -59,23 +62,29 @@ int main(void) {
         return  EXIT_FAILURE;
     }
 
-// check for a NULL query string -------------------------------------------------------------------------------------=
+// check for a NULL query string ---------------------------------------------------------------------------------------
 
-//    setenv("QUERY_STRING", "titleID=147", 1);
+//    setenv("QUERY_STRING", "titleID=147", 1);                                            // uncomment for testing only
 
     sParam = getenv("QUERY_STRING");
 
     if(sParam == NULL) {
-        printf("Query string is empty. Terminating program");
+        printf("Query string is NULL. Expecting QUERY_STRING=\"titleID=<999>\". Terminating bookDelBook.cgi");
         printf("\n\n");
         return EXIT_FAILURE;
     }
 
-//  get the content from QUERY_STRING and tokenize the ratingID value -------------------------------------------------
+//  get the Title ID from QUERY_STRING ---------------------------------------------------------------------------------
 
     sscanf(sParam, "titleID=%d", &iTitleID);
 
-// set a SQL query to insert the new author ---------------------------------------------------------------------------
+    if(iTitleID == 0) {
+        printf("Title ID string is 0. Expecting QUERY_STRING=\"titleID=<999>\". Terminating bookDelBook.cgi");
+        printf("\n\n");
+        return EXIT_FAILURE;
+    }
+
+// set a SQL query to insert the new author ----------------------------------------------------------------------------
 
     sprintf(caSQL, "DELETE FROM risingfast.`Book Titles` "
                    "WHERE `Title ID` = %d;", iTitleID);
