+++ bookDelRating.c
@@ -7,6 +7,9 @@
  *      09-Dec-2021 start by copying bookDelClassification.c
  *      13-Dec-2021 check if no rating deleted
  *      15-Sep-2022 add Access-Control-Allow-Origin: * CORS http header
+ *      11-Oct-2022 clean up comments
+ *      11-Oct-2022 use EXIT_SUCCESS and EXIT_FAILURE on returns
+ *      11-Oct-2022 validate QUERY_STRING for NULL and empty values
  *  Enhancements:
  */
 #include <mysql.h>
@@ -21,12 +24,12 @@
 
 #define MAXLEN 1024
 
-// global declarations
+// global declarations -------------------------------------------------------------------------------------------------
 
-char *sgServer = "192.168.0.13";                                                               //mysqlServer IP address
-char *sgUsername = "gjarman";                                                              // mysqlSerer logon username
-char *sgPassword = "Mpa4egu$";                                                    // password to connect to mysqlserver
-char *sgDatabase = "risingfast";                                                // default database name on mysqlserver
+char *sgServer = "192.168.0.13";                                                                //mysqlServer IP address
+char *sgUsername = "gjarman";                                                               // mysqlSerer logon username
+char *sgPassword = "Mpa4egu$";                                                     // password to connect to mysqlserver
+char *sgDatabase = "risingfast";                                                 // default database name on mysqlserver
 
 MYSQL *conn;
 MYSQL_RES *res;
@@ -41,12 +44,12 @@ int main(void) {
 
     char caSQL[SQL_LEN] = {'\0'};
 
-// print the html content type and CORS <header> block ----------------------------------------------------------------
+// print the html content type and CORS <header> block -----------------------------------------------------------------
 
     printf("Content-type: text/html\n");
     printf("Access-Control-Allow-Origin: *\n\n");
 
-// Initialize a connection and connect to the database
+// Initialize a connection and connect to the database -----------------------------------------------------------------
 
     conn = mysql_init(NULL);
 
@@ -59,23 +62,36 @@ int main(void) {
         return  EXIT_FAILURE;
     }
 
-// check for a NULL query string --------------------------------------------------------------------------------------
+//    setenv("QUERY_STRING", "ratingID=22", 1);                                            // uncomment for testing only
 
-//    setenv("QUERY_STRING", "ratingID=22", 1);
+// check for a NULL query string ---------------------------------------------------------------------------------------
 
     sParam = getenv("QUERY_STRING");
 
     if(sParam == NULL) {
-        printf("Query string is empty. Terminating program");
+        printf("Query string is NULL. Expecting QUERY_STRING=\"ratingID=<99>\". Terminating bookDelRating.cgi");
         printf("\n\n");
         return EXIT_FAILURE;
     }
 
-//  get the content from QUERY_STRING and tokenize the ratingID value -------------------------------------------------
+// check for a empty (non-NULL) query string ---------------------------------------------------------------------------------------
+
+    if(sParam == NULL) {
+        printf("Query string is empty (non-NULL). Expecting QUERY_STRING=\"ratingID=<99>\". Terminating bookDelRating.cgi");
+        printf("\n\n");
+        return EXIT_FAILURE;
+    }
+
+//  get the content from QUERY_STRING and tokenize the ratingID value --------------------------------------------------
 
     sscanf(sParam, "ratingID=%d", &iRatingID);
+    if(iRatingID == 0) {
+        printf("Rating ID is 0. Expecting QUERY_STRING=\"ratingID=<99>\". Terminating bookDelRating.cgi");
+        printf("\n\n");
+        return EXIT_FAILURE;
+    }
 
-// set a SQL query to insert the new author ---------------------------------------------------------------------------
+// set a SQL query to insert the new author ----------------------------------------------------------------------------
 
     sprintf(caSQL, "DELETE FROM risingfast.`Book Ratings` "
                    "WHERE `Rating ID` = %d;", iRatingID);
